third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/multicol-fragmented-anchor-expected.txt

This is a testharness.js-based test.
[FAIL] An element in a fragmented div should be able to be selected as an anchor node.
  assert_equals: expected 20 but got 10
Harness: the test ran to completion.
